websocket-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-socketio==5.12.1
python-engineio
psycopg[binary]==3.1.18
python-dotenv==1.0.1
pydantic>=2.5.0
prometheus-client==0.19.0
asyncpg==0.29.0
gunicorn==21.2.0
aiohttp==3.9.3 